Added TryCatch decorator and Database

+++ examples/test-app/scripts/clean.ps1
@@ -1,3 +1,8 @@
-Remove-Item -Recurse .\node_modules -Force
 
-Remove-Item .\package-lock.json -Force
+if (Test-Path -Path .\node_modules) {
+    Remove-Item -Recurse .\node_modules -Force
+}
+
+if (Test-Path -Path .\package-lock.json) {
+    Remove-Item .\package-lock.json -Force
+}
